rename hooks to match codebase convention
