+++ frontend/src/metabase/dashboard/components/DashboardTabs/DashboardTabs.unit.spec.tsx
@@ -9,7 +9,7 @@ import { getDefaultTab, resetTempTabId } from "metabase/dashboard/actions";
 import { INPUT_WRAPPER_TEST_ID } from "metabase/core/components/TabButton";
 import { DashboardTabs } from "./DashboardTabs";
 import { TEST_DASHBOARD_STATE } from "./test-utils";
-import { useDashboardTabs } from "./useDashboardTabs";
+import { useDashboardTabs } from "./use-dashboard-tabs";
 
 function setup({
   isEditing = true,
