+++ frontend/src/metabase/dashboard/components/DashboardTabs/use-sync-url-slug.ts
